Feature: pengine: formatted output for log
This PR complements the PRs #1834, #1821 and #1788. It adds other 4 functions-messages
 to print information about resources in HTML by splitting/refactoring

    native_print --> pe__resource_log
    group_print --> pe__group_log
    clone_print --> pe__clone_log
    pe__print_bundle --> pe__bundle_log

Those three functions require the integer log_level parameter.
It cannot be the third parameter, because the third parameter
is const char *pre_text. So I made it be the fourth parameter.
The pre_text is also used, but I think, after carefull refactoring
both, pre_text and log_level should go.
Like the aforementioned PRs, this one adds the supplementary functions:

    print_bundle_replica --> pe__bundle_replica_output_log
    common_print --> pe__common_output_log
    print_rscs_brief --> pe__rscs_brief_output_log
    short_print --> pe__short_output_log
    native_print_attr --> pe__native_output_attr_log

I see two directions of refactoring now:
1) make a good log formatter (implemented in the previous commit)
so that you could use the *_text functions instead of the *_log ones.
2) make a good formatted and refactor the *_log functions in this commit
to avoid using the pre_text and log_level in the *_log functions.

TODO:

1) read what does the Ken say, what should I exactly do

2) create a macro like

  pcmk__new_log_message(level, __FILE__, __func__, __LINE__, fmt, ##args)

where the pcmk__new_log_message is the function that returns void*
and declare in the lib/common/output_log.c like

void* pcmk__new_log_message(int level
  , const char *file
  , const char *function
  , int line
  , const char *fmt
  , args...)
{
  char text[LINE_MAX];
  va_start(arg);
  // look how you do in in the ASSERT macro in the interpreter
  vfprintf(text, LINE_MAX, fmt, ##args);
  va_stop(arg);
  private_log_message_t* result
    = calloc(1, sizeof(private_log_message_t));
  result->level = level;
  result->file = str_dup(file);
  result->function = str_dup(function);
  result->line = line;
  result->text = str_dup(text);
  return result;
}

TAKE INTO ACCOUNT: maybe you don't even have to make it that complex,
because the user expects to see the logs immediatelly as they come.
In this case you can simply leave the print_status as it was, haha.

On one hand, it would contradict the idea to gather the information and deside
in which format to print it in the end.

On the other hand you do it for the *_text case, where you immediatelly
call printf(out->dest, ...)