Feature: pengine: formatted output for log
This PR complements the PRs #1834, #1821 and #1788. It adds other 4 functions-messages
 to print information about resources in HTML by splitting/refactoring

    native_print --> pe__resource_log
    group_print --> pe__group_log
    clone_print --> pe__clone_log
    pe__print_bundle --> pe__bundle_log

Those three functions require the integer log_level parameter.
It cannot be the third parameter, because the third parameter
is const char *pre_text. So I made it be the fourth parameter.
The pre_text is also used, but I think, after carefull refactoring
both, pre_text and log_level should go.
Like the aforementioned PRs, this one adds the supplementary functions:

    print_bundle_replica --> pe__bundle_replica_output_log
    common_print --> pe__common_output_log
    print_rscs_brief --> pe__rscs_brief_output_log
    short_print --> pe__short_output_log
    native_print_attr --> pe__native_output_attr_log