Dev: mylog
There are so many log functions in pacemaker. It's easier to
implement my own log function rather than understand how to use those.
Here I add mylog function and an example on how to use it (in
ipc_controld.c).
To test it, for examlpe, call $ crmadmin -D
This would write to the ~hacluster/crmadmin.txt
'''
11:20:44, crmadmin(32181), ipc_controld.c:dispatch:250, MSG:value=ping
'''