Explanation how the ipc works in crmadmin and pacemaker-controld
The interraction between crmadmin and pacemaker-controld is done in
xml format. When you do 'crmadmin -D' the sequence is as follows:
1) crmadmin::pcmk__send_ipc_request::crm_ipc_send --> controld::route_message::relay_message
2) controld::route_message::relay_message::send_msg_via_ipc --> crmadmin::pcmk__send_ipc_request

Besides, there are two other commits in here

1) The mylog + mylog_xml functionality

2) Fix: controld: leave xml-src attribute empty when no DC selected #2902

( Right after starting the cluster the DC is offline for short time,
and if you call 'crmadmin --dc_lookup' during this time, it would hang.
This change leaves the xml-src attribute that the crmadmin receives from
the pacemaker-controld empty, so that crmadmin knows the DC was not
selected.
ref: https://projects.clusterlabs.org/T735 )